docshell/test/chrome/test_open_and_immediately_close_opener.html

<!DOCTYPE html>
<html>
<head>
  <title>Test for Bug 1702678</title>
  <meta charset="utf-8">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1702678">Mozilla Bug 1702678</a>

<script type="application/javascript">
"use strict";

const HTML = `
<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <script>
    // We need to queue a promise reaction job whilch will close the window
    // during the nested event loop spun up by the window opening code.
    Promise.resolve().then(() => {
      window.close();
    });
    window.open("data:text/html,Hello");
  <\/script>
</head>
</html>
`;

add_task(async function() {
  // This bug only manifests when opening tabs in new windows.
  await SpecialPowers.pushPrefEnv({
    set: [["browser.link.open_newwindow", 2]],
  });

  // Create a window in a new BrowsingContextGroup so that it will be the last
  // window in the group when it closes, and the group will be destroyed.
  window.open(`data:text/html,${encodeURIComponent(HTML)}`, "", "noopener");

  // Make a few trips through the event loop to ensure we've had a chance to
  // open and close the relevant windows.
  for (let i = 0; i < 10; i++) {
    await new Promise(resolve => setTimeout(resolve, 0));
  }

  ok(true, "We didn't crash");
});
</script>

</body>
</html>
